FEAT: Add Exception Handler

+++ src/main/java/com/ktds/dsquare/board/card/CardService.java
@@ -6,6 +6,8 @@ import com.ktds.dsquare.board.comment.CommentService;
 import com.ktds.dsquare.board.enums.BoardType;
 import com.ktds.dsquare.board.like.LikeRepository;
 import com.ktds.dsquare.board.paging.PagingService;
+import com.ktds.dsquare.common.exception.EntityNotFoundException;
+import com.ktds.dsquare.common.exception.PostNotFoundException;
 import com.ktds.dsquare.member.Member;
 import com.ktds.dsquare.member.dto.response.BriefMemberInfo;
 import com.ktds.dsquare.member.dto.response.TeamInfo;
@@ -16,7 +18,6 @@ import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 
-import javax.persistence.EntityNotFoundException;
 import javax.transaction.Transactional;
 import java.util.ArrayList;
 import java.util.List;
@@ -96,6 +97,8 @@ public class CardService {
     //read - 카드주세요 글 상세 조회
     public CardResponse getCardDetail(Long cardId, Member user) {
         Card card = cardRepository.findByDeleteYnAndId(false, cardId);
+        if(card == null)
+            throw new PostNotFoundException("card not found. Card ID: "+cardId);
         card.increaseViewCnt();
         cardRepository.save(card);
 
@@ -108,6 +111,8 @@ public class CardService {
     @Transactional
     public void giveCard(Long cardId, Member user){
         Card card = cardRepository.findByDeleteYnAndId(false, cardId);
+        if(card == null)
+            throw new PostNotFoundException("card not found. Card ID: "+cardId);
         card.selectCard(user, true);
     }
 
@@ -115,6 +120,8 @@ public class CardService {
     @Transactional
     public void updateCard(Long cardId, CardUpdateRequest request){
         Card card = cardRepository.findByDeleteYnAndId(false, cardId);
+        if(card == null)
+            throw new PostNotFoundException("card not found. Card ID: "+cardId);
         Team projTeam = teamRepository.findById(request.getProjTeamId())
                 .orElseThrow(() -> new EntityNotFoundException("team is not found"));
 
@@ -124,8 +131,9 @@ public class CardService {
     //delete - 카드주세요 글 삭제
     @Transactional
     public void deleteCard(Long cardId){
-        Card card = cardRepository.findById(cardId)
-                .orElseThrow(()-> new EntityNotFoundException("card is not found"));
+        Card card = cardRepository.findByDeleteYnAndId(false, cardId);
+        if(card == null)
+            throw new PostNotFoundException("card not found. Card ID: "+cardId);
         card.deleteCard();
         commentService.deleteCommentCascade(BoardType.CARD, cardId);
     }
@@ -153,15 +161,17 @@ public class CardService {
     }
 
     public void like(Long id) {
-        Card card = cardRepository.findById(id)
-                .orElseThrow(()-> new EntityNotFoundException("answer not found"));
+        Card card = cardRepository.findByDeleteYnAndId(false, id);
+        if(card == null)
+            throw new PostNotFoundException("card not found. Card ID: "+id);
         card.like();
     }
 
 
     public void cancleLike(Long id){
-        Card card = cardRepository.findById(id)
-                .orElseThrow(()-> new EntityNotFoundException("answer not found"));
+        Card card = cardRepository.findByDeleteYnAndId(false, id);
+        if(card == null)
+            throw new PostNotFoundException("card not found. Card ID: "+id);
         card.cancleLike();
     }
